# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a high-performance dating app backend server.",
        "title": "Dating App API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches a list of chat conversations with existing messages for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get user conversations (Active)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor (RFC3339 time format)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversations list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/chat.ConversationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chat/conversations/match/{matchId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the conversation associated with a specific match ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get conversation by match ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Match ID",
                        "name": "matchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/chat.ConversationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chat/conversations/{id}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the message history for a specific conversation with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get conversation messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message history",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/chat.MessageResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chat/new-matches": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of conversations with no messages, intended for the \"New Matches\" horizontal row.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get new matches (no messages)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit (default 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Cursor (RFC3339 time format)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New matches list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/chat.ConversationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/chat/upload-url": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Provides a temporary upload URL for media attachments in chat.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get chat media upload URL",
                "responses": {
                    "501": {
                        "description": "Not implemented",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/entities/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the complete profile details of an entity (solo user or group) by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "entities"
                ],
                "summary": "Get entity details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Entity ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity profile",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.EntityResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "Entity not found",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chat.ConversationResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "avatar_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "entity": {
                    "$ref": "#/definitions/response.EntityResponse"
                },
                "id": {
                    "type": "string"
                },
                "is_typing": {
                    "type": "boolean"
                },
                "last_message": {
                    "$ref": "#/definitions/chat.MessageResponse"
                },
                "swiper_entity_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "\"direct\" or \"group\"",
                    "type": "string"
                },
                "unread_count": {
                    "type": "integer"
                }
            }
        },
        "chat.MessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "metadata": {
                    "$ref": "#/definitions/entities.MessageMetadata"
                },
                "sender_id": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entities.MessageType"
                }
            }
        },
...
        "entities.MessageMetadata": {
            "type": "object",
            "properties": {
                "gif_provider": {
                    "type": "string"
                },
                "image_height": {
                    "type": "integer"
                },
                "image_width": {
                    "type": "integer"
                }
            }
        },
        "entities.MessageType": {
            "type": "string",
            "enum": [
                "text",
                "image",
                "gif"
            ],
            "x-enum-varnames": [
                "MessageTypeText",
                "MessageTypeImage",
                "MessageTypeGif"
            ]
        },
...
        "entities.UserStatus": {
            "type": "string",
            "enum": [
                "onboarding",
                "active",
                "banned"
            ],
            "x-enum-varnames": [
                "UserStatusOnboarding",
                "UserStatusActive",
                "UserStatusBanned"
            ]
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "response.ConsumableItemResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "item_type": {
                    "type": "string"
                }
            }
        },
        "response.EntityResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/response.GroupResponse"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/response.UserResponse"
                }
            }
        },
        "response.GroupResponse": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "main_photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.UserResponse"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.MasterItemResponse": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.PhotoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_main": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "response.UserResponse": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "bio": {
                    "type": "string"
                },
                "consumables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.ConsumableItemResponse"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "height_cm": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "interested_genders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.MasterItemResponse"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "location_city": {
                    "type": "string"
                },
                "location_country": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "main_photo": {
                    "type": "string"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PhotoResponse"
                    }
                },
                "relationship_type": {
                    "$ref": "#/definitions/response.MasterItemResponse"
                },
                "status": {
                    "$ref": "#/definitions/entities.UserStatus"
                },
                "subscription": {
                    "$ref": "#/definitions/response.UserSubscriptionResponse"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
        "response.UserSubscriptionResponse": {
            "type": "object",
            "properties": {
                "expired_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "plan_id": {
                    "type": "string"
                },
                "plan_name": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer ' followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}